refactor: tripShareButton 컴포넌트 분리
